requirements.txt
gradio 
flask==2.0.1
flask-cors==3.0.10
pandas==1.3.3
openpyxl==3.0.9 

# web / api
fastapi>=0.95,<1.0
starlette>=0.27,<1.0
uvicorn>=0.20,<1.0

# db / http
pymysql>=1.0,<2.0
requests>=2.28,<3.0

# llm sdk
openai>=1.0,<2.0

# image / compute
numpy>=1.19,<2.0
opencv-python>=4.7,<5.0
Pillow>=9.0,<11.0
pdf2image>=1.16,<2.0

# segmentation / detection
ultralytics>=8.0,<9.0
segment-anything>=1.0,<2.0
pycocotools>=2.0.6,<3.0

# deep learning
torch>=2.0,<3.0
torchvision>=0.15,<1.0

# ocr
paddleocr>=2.7,<3.0
paddlepaddle>=2.5,<3.0